Restructure preview component
